Extract server import zip helpers
